[M1] refactor: 패키지 경로 정리
https://github.com/next-step/java-lotto/pull/2508
