+++ src/test/java/calculator/domain/SingleFormulaTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package calculator.domain;
 
+import calculator.domain.SingleFormula;
+import calculator.domain.SingleFormulaVo;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
